requirements.txt
flask
langchain
langchain-community
pypdf
python-docx
pandas
openpyxl
faiss-cpu